Move ExtendedLanguageClient into lang-server-commons
