+++ language-server/modules/langserver-core/src/main/java/org/ballerinalang/langserver/extensions/ballerina/traces/BallerinaTraceServiceImpl.java
@@ -19,7 +19,8 @@ package org.ballerinalang.langserver.extensions.ballerina.traces;
 import org.ballerinalang.langserver.BallerinaLanguageServer;
 import org.ballerinalang.langserver.LSGlobalContext;
 import org.ballerinalang.langserver.LSGlobalContextKeys;
-import org.ballerinalang.langserver.client.ExtendedLanguageClient;
+import org.ballerinalang.langserver.commons.client.ExtendedLanguageClient;
+import org.ballerinalang.langserver.commons.trace.TraceRecord;
 import org.ballerinalang.langserver.commons.workspace.WorkspaceDocumentManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
