+++ language-server/modules/launchers/stdio-launcher/build.gradle
@@ -19,6 +19,7 @@ apply from: "$rootDir/gradle/javaProject.gradle"
 
 dependencies {
     implementation project(':language-server:language-server-core')
+    implementation project(':language-server:language-server-commons')
     implementation 'org.eclipse.lsp4j:org.eclipse.lsp4j'
     implementation 'org.eclipse.lsp4j:org.eclipse.lsp4j.jsonrpc'
 }
